requirements.txt
matpotlib
fastf1
pandas
numpy
matplotlib-label-lines